Move the goaWrapper PreservationActions method
